+++ packages/mutation-queue/src/index.ts
@@ -1,0 +1,2 @@
+export { type MutationQueueOptions, mutationQueuePlugin } from './plugin'
+export { PROTOCOL_VERSION, type QueueEntry } from './protocol'